fix: move blob factor to its own testutil package
